Partial fix to HARMONY-3197 [drlvm][jit][winx64] Jitrino compilation fixed`. Tested on SUSE10@ia32, SUSE9@x64, WinXP@ia32, Win2003@x64, SUSE9@ipf
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@512698 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/jitrino/src/codegenerator/ia32/Ia32ConstraintsResolver.cpp
@@ -384,7 +384,7 @@ void ConstraintsResolverImpl::createBasicBlockArray()
         if (node->isBlockNode()){
             double bbecv = getBasicBlockPriority(node);
             uint32 ibb=0;
-            for (uint32 nbb=basicBlocks.size(); ibb<nbb; ++ibb){
+            for (uint32 nbb=(uint32)basicBlocks.size(); ibb<nbb; ++ibb){
                 if (bbecv > getBasicBlockPriority(basicBlocks[ibb]))
                     break;
             }
@@ -671,7 +671,7 @@ void ConstraintsResolverImpl::resolveConstraints(Node * bb)
 void ConstraintsResolverImpl::resolveConstraints()
 {   
     // for all basic blocks in the array
-    for (uint32 ibb=0, nbb=basicBlocks.size(); ibb<nbb; ++ibb){
+    for (uint32 ibb=0, nbb=(uint32)basicBlocks.size(); ibb<nbb; ++ibb){
         resolveConstraints(basicBlocks[ibb]);
     }
 }
